+++ pro-admin/src/main/java/dr/sbs/admin/service/AdminMenuService.java
@@ -11,20 +11,20 @@ public interface AdminMenuService {
   boolean create(AdminMenu adminMenu);
 
   /** 修改后台菜单 */
-  boolean update(Long id, AdminMenu adminMenu);
+  boolean update(Integer id, AdminMenu adminMenu);
 
   /** 根据ID获取菜单详情 */
-  AdminMenu getItem(Long id);
+  AdminMenu getItem(Integer id);
 
   /** 根据ID删除菜单 */
-  boolean delete(Long id);
+  boolean delete(Integer id);
 
   /** 分页查询后台菜单 */
-  Page<AdminMenu> list(Long parentId, Integer pageSize, Integer pageNum);
+  Page<AdminMenu> list(Integer parentId, Integer pageSize, Integer pageNum);
 
   /** 树形结构返回所有菜单列表 */
   List<AdminMenuNode> treeList();
 
   /** 修改菜单显示状态 */
-  boolean updateHidden(Long id, Integer hidden);
+  boolean updateHidden(Integer id, Integer hidden);
 }
