+++ pro-admin/src/main/java/dr/sbs/admin/service/AdminUserCacheService.java
@@ -7,19 +7,19 @@ import java.util.List;
 /** 后台用户缓存操作类 */
 public interface AdminUserCacheService {
   /** 删除后台用户缓存 */
-  void delUser(Long userId);
+  void delUser(Integer userId);
 
   /** 删除后台用户资源列表缓存 */
-  void delResourceList(Long userId);
+  void delResourceList(Integer userId);
 
   /** 当角色相关资源信息改变时删除相关后台用户缓存 */
-  void delResourceListByRole(Long roleId);
+  void delResourceListByRole(Integer roleId);
 
   /** 当角色相关资源信息改变时删除相关后台用户缓存 */
-  void delResourceListByRoleIds(List<Long> roleIds);
+  void delResourceListByRoleIds(List<Integer> roleIds);
 
   /** 当资源信息改变时，删除资源项目后台用户缓存 */
-  void delResourceListByResource(Long resourceId);
+  void delResourceListByResource(Integer resourceId);
 
   /** 获取缓存后台用户信息 */
   AdminUser getUser(String username);
@@ -28,8 +28,8 @@ public interface AdminUserCacheService {
   void setUser(AdminUser adminUser);
 
   /** 获取缓存后台用户资源列表 */
-  List<AdminResource> getResourceList(Long userId);
+  List<AdminResource> getResourceList(Integer userId);
 
   /** 设置后台后台用户资源列表 */
-  void setResourceList(Long userId, List<AdminResource> resourceList);
+  void setResourceList(Integer userId, List<AdminResource> resourceList);
 }
