+++ pro-mp/src/main/java/dr/sbs/mp/entity/AdminMenu.java
@@ -27,10 +27,10 @@ public class AdminMenu implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     @ApiModelProperty("父级ID")
-    private Long parentId;
+    private Integer parentId;
 
     @ApiModelProperty("菜单名称")
     private String title;
